// apps/web/src/features/auth/components/LoginForm/LoginForm.types.ts
import * as z from "zod";

export const loginSchema = z.object({
	email: z.string().email("Ingrese un correo corporativo válido"),
	password: z.string().min(6, "La contraseña debe tener al menos 6 caracteres"),
	rememberMe: z.boolean().optional(),
});

export type LoginFormData = z.infer<typeof loginSchema>;